+++ trailing_zeros_19.go
@@ -4,6 +4,6 @@ package bitset
 
 import "math/bits"
 
-func trailingZeroes64(v uint64) uint {
-	return uint(bits.TrailingZeros64(v))
+func trailingZeroes64(v uint64) uint64 {
+	return uint64(bits.TrailingZeros64(v))
 }
